Sistemati alcuni bug del viewer e delle simulazioni. Rimosso il Menù e lasciato solamente la toolbar. Rimossa l'action new_item per creare un bottone sopra search per l'aggiunta di un nuovo sensore. Aggiunte immagini per molti elementi

+++ Core/Activity.cpp
@@ -3,7 +3,7 @@
 namespace Core
 {
     Activity::Activity(unsigned int id, std::string name, std::string description, unsigned int age, float height, float weight, unsigned int training_type,
-    float training_time, unsigned int bpm, float calories, float speed, float distance) :
+    float training_time, unsigned int bpm, unsigned int calories, float speed, float distance) :
     Sensor(id, name, description, age, height, weight, training_type, training_time), 
     CaloriesCounter(id, name, description, age, height, weight, training_type, training_time, bpm, calories),
     Speedometer(id, name, description, age, height, weight, training_type, training_time, speed, distance) {}
